requirements.txt
google-api-python-client==2.108.0
google-auth==2.23.4
google-generativeai==0.3.1
python-dotenv==1.0.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pandas==2.1.3
